backend/requirements.txt
flask
flask-socketio
flask-sock
flask-cors
flask-pymongo
pymongo
paho-mqtt
python-dotenv
eventLet
python-engineio
python-socketio